+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [2.0.0 - Unreleased]
+### Removed
+- Remove `.getClient()` method
+### Added
+- Add `.rates()` method for exchange rates
+- Add `.history()` method for historical rates
+- Add Standard.js for linting
+- Add Github action to publish to NPM Registry and Github Packages
+
 ## [1.1.0]
 ### Fixed
 - Fix API base url from `https://exchangeratesapi.io/api` to `https://api.exchangeratesapi.io`